+++ core/board.py
@@ -2,8 +2,8 @@ class Board:
     def __init__(self):
 
         self.__columnas__ = [{'checker' : 'x', 'quantity' : 9}, {'checker' : 'x', 'quantity' : 3}, {'checker' : 'x', 'quantity' : 5}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0},
-                             {'checker' : 'o', 'quantity' : 3}, {'checker' : 'o', 'quantity' : 4}, {'checker' : 'o', 'quantity' : 4}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}]
-        self.__barra__ = ()
+                             {'checker' : 'o', 'quantity' : 3}, {'checker' : 'o', 'quantity' : 4}, {'checker' : 'o', 'quantity' : 4}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0}, {'checker' : '', 'quantity' : 0},
+                             {'bar': True, 'checker' : 'x', 'quantity' : 0}, {'bar': True, 'checker' : 'o', 'quantity' : 0}]
     def show_board(self):
         """ Este método imprime el tablero, poniendo el método checker(column, level) con las respectivas coordenadas en cada espacio ocupable por fichas del tablero."""
         c = self.checker
@@ -14,7 +14,8 @@ class Board:
         for y in range(1, 6):
             left = ' | '.join([f"{c(x,y)}" for x in range(12, 18, 1)])
             right = ' | '.join([f"{c(x,y)}" for x in range(18, 24, 1)])
-            print(f"Fila {y}   | {left} |  |   |  | {right} |")
+            bar_o = c(25, y)
+            print(f"Fila {y}   | {left} |  | {bar_o} |  | {right} |")
         print("Fila 6   | v | v | v | v | v | v |  |   |  | v | v | v | v | v | v |")
         print("         +---+---+---+---+---+---+  |   |  +---+---+---+---+---+---+")
         print("         +---+---+---+---+---+---+  |   |  +---+---+---+---+---+---+")
@@ -22,14 +23,15 @@ class Board:
         for y in range(5, 0, -1):
             left = ' | '.join([f"{c(x,y)}" for x in range(11, 5, -1)])
             right = ' | '.join([f"{c(x,y)}" for x in range(5, -1, -1)])
-            print(f"Fila {y}   | {left} |  |   |  | {right} |")
+            bar_x = c(24, y)
+            print(f"Fila {y}   | {left} |  | {bar_x} |  | {right} |")
         print("         +---+---+---+---+---+---+  |   |  +---+---+---+---+---+---+")
         print("          12  11  10   9   8   7    |BAR|    6   5   4   3   2   1")
         print("- La columna central es la **BAR** (barra).")
         print("- Si un punto tiene más de 5 fichas, usa (6), (7), etc.")
     
     def checker(self, column:int, level:int):
-        """ Este método devuelve la ficha correspondiente (o ninguna) dependiendo de las coordenadas del tablero y la ocupacion, que analiza en el atributo __columnas__. """
+        """ Este método devuelve la ficha correspondiente (o ninguna) dependiendo de las coordenadas del tablero y la ocupacion, que analiza en el atributo __columnas__ y __barra__ (para fichas en las barras)  """
         q = self.__columnas__[column]['quantity']
         if  q == 0:
             return ' '
@@ -39,7 +41,7 @@ class Board:
             return self.__columnas__[column]["checker"]
         else:
             return ' '
-        
+            
     def add_checker(self, column:int, quan = 1):
         """ Este método aumenta por 'quan' unidades (1 por defecto) la cantidad de fichas que se encuentran en una columna del tablero, accediendo al atributo __columnas__. Debe haberse configurado la ficha previamente para que su uso tenga sentido. """
         self.__columnas__[column]['quantity'] += quan
@@ -61,4 +63,5 @@ class Board:
         """ Este método devuelve el atributo __columnas__ para su uso externamente. """
         return self.__columnas__
 
-
+# b = Board()
+# b.show_board()